+++ app/javascript/stylesheets/sitewide/color-library.scss
@@ -6,18 +6,21 @@
     --plum-hue-sat:  265, 85%;
     --peacock-hue-sat:  217, 100%;
 
-    --site-background: var(--color-fill-inverse);
+    --site-background-hue-sat: 0, 0%;
+    --site-background-lightness: 10%;
+    --site-background-adjusted-lightness:calc(var(--site-background-lightness) + (25% * var(--color-lightness-shift-basis)));
+    --site-background: hsl(var(--site-background-hue-sat), var(--site-background-adjusted-lightness));
 
     --inverse-hue-sat:  0, 0%;
-    --inverse-lightness: 100%;
+    --inverse-lightness: 95%;
 
     --neutral-hue-sat: 0, 0%;
-    --neutral-lightness: 0%;
+    --neutral-lightness: 5%;
 
     --saturated-fill-lightness: 25%;
-    --color-shift-basis: 1;
+    --saturated-text-lightness: calc(8% * var(--color-lightness-shift-basis));
+    --color-lightness-shift-basis: 1;
     --desaturated-hue-sat: 0, 0%;
-
     //Derivative values 
 
     --color-fill-inverse-hue-sat: var(--inverse-hue-sat);
@@ -36,17 +39,17 @@
     --color-fill-theme-highlight: hsl(var(--color-fill-theme-hue-sat), var(--color-fill-theme-highlight-lightness)); 
 
     --color-fill-positive-hue-sat: 168, 76%;
-    --color-fill-positive-lightness:var(--saturated-fill-lightness);
+    --color-fill-positive-lightness: 25%;
     --color-fill-positive: hsl(var(--color-fill-positive-hue-sat),var(--color-fill-positive-lightness));
     
     --color-fill-negative-hue-sat: 2, 100%;
     // currently negative is a red and reds need a slightly higher lightness to come through 
-    --color-fill-negative-lightness: calc(var(--saturated-fill-lightness) + 20%);
+    --color-fill-negative-lightness: 50%;
     --color-fill-negative: hsl(var(--color-fill-negative-hue-sat),var(--color-fill-negative-lightness));
     
     --color-fill-alert-hue-sat: 25, 100%;
     // currently alert is a orange and reds need a slightly higher lightness to come through 
-    --color-fill-alert-lightness: calc(var(--saturated-fill-lightness) + 25%);
+    --color-fill-alert-lightness: 50%;
     --color-fill-alert: hsl(var(--color-fill-alert-hue-sat),var(--color-fill-alert-lightness));
     
 
@@ -59,7 +62,7 @@
     --color-text-neutral: var(--color-fill-inverse); 
 
     --color-text-theme-hue-sat: var(--color-fill-theme-hue-sat);
-    --color-text-theme-lightness: 10%;
+    --color-text-theme-lightness: var(--saturated-text-lightness);
     --color-text-theme: hsl(var(--color-text-theme-hue-sat), var(--color-text-theme-lightness));
